Moved DataclassFactoryHandler, fixed handlers for polar and bikecitizens
